backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
langgraph==0.2.61
langchain-openai==0.2.14
python-dotenv==1.0.1
pinecone-plugin-interface==0.0.7
pinecone==5.4.2
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
# Phase 1: Database + Auth
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
# Phase 3: Rate limiting
slowapi>=0.1.9
